pyproject.toml
[project]
name = "auto-dev-orchestrator-py"
version = "0.1.0"
description = "Production-ready orchestration service for automated software delivery"
authors = [{ name = "AutoDev AI", email = "ops@example.com" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "anyio>=4.4",
    "celery>=5.4",
    "fastapi>=0.115",
    "pydantic>=2.7",
    "GitPython>=3.1.44",
    "gradio>=4.44",
    "httpx>=0.27",
    "PyGithub>=2.3",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.4",
    "requests>=2.32",
    "sqlalchemy>=2.0.34",
    "structlog>=24.1",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
tests = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
testpaths = ["tests"]
norecursedirs = ["webui"]
